Initial import of site generator and content
Content is either written as plain HTMLs, Mustache templates or Markdown
files. The top-level files are considered as pages. The items under
"posts/" are considered blog posts. Note that this can be configured
in the "site.json" file.

A lot of things are in flux and are *very* likely to change -- mostly
to suit my personal needs.

The site generator (site.js + build) sues Node.JS, plus the following
modules:

- veil (loading metadata and content bodies, rfc822-style).
- marked (renders GitHub-flavored Markdown).
- mustache (used for templating).
- glob (used for determining which files are to be rendered).
- classy.js (included, see classy.pocoo.org)