layout: Remove setting the base href
This causes some troubles when loading the page through
www.perezdecastro.org (e.g. CSS style sheets may not load) due to
CSP headers not including the www.* version. The <base> is not really
needed now that Disqus comments are not there anymore, so let's slash
the corresponding HTML snippet.